Add pinhole and fisheye camera model calibration
Remove the older calib_find_intrinsics.py to make two
different ones. calib_find_intrinsic_video.py which is
essentially the same as calib_find_intrinsics.py but with
a few updates. With video, it is hard to "intelligently"
grab those frames that are necessary for calibration.
So, the move calib_find_intrinsics_images.py. Now,
images that are possibly good for calibration are
selected using calib_save_images.py and then reread
in the aforementioned script.